+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Added support of `datetime` type for Tarantool 2.10+
+
 ## [1.6.0] - 2022-02-01
 
 ### Added